add delete folder and files

+++ routes/folders.js
@@ -1,55 +1,79 @@
-const expires = require("express");
-const {PrismaClient} = require("@prisma/client");
-const router = expires.Router();
+const express = require("express");
+const { PrismaClient } = require("@prisma/client");
+const router = express.Router();
 const prisma = new PrismaClient();
+const ensureAuthenticated = require("./auth");
 
-router.post("/", async (req, res) => {
+router.post("/", ensureAuthenticated, async (req, res) => {
   const { name } = req.body;
   try {
     const folder = await prisma.folder.create({
-      data: {name},
+      data: { name, userId: req.user.id,
+      },
     });
     res.status(201).json(folder);
-  } catch (error){
-    res.status(500).json({error: "Error creating folder"});
+  } catch (error) {
+    res.status(500).json({ error: "Error creating folder" });
   }
-})
+});
 
-router.get("/", async (req, res) => {
+router.get("/", ensureAuthenticated, async (req, res) => {
   try {
     const folders = await prisma.folder.findMany({
-      include: {files: true},
+      where: { userId: req.user.id},
+      include: { files: true },
     });
     res.json(folders);
-  } catch (error){
-    res.status(500).json({ error: "Error fetching folders"});
+  } catch (error) {
+    res.status(500).json({ error: "Error fetching folders" });
   }
-})
+});
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", ensureAuthenticated, async (req, res) => {
   const { id } = req.params;
   const { name } = req.body;
-  try{
+  try {
     const folder = await prisma.folder.update({
       where: { id: parseInt(id) },
-      data: {name},
+      data: { name },
     });
     res.json(folder);
-  } catch (error){
+  } catch (error) {
     res.status(500).json({ error: "Error updating folder" });
   }
-})
+});
 
-router.delete("/:id", async (req, res) => {
+router.delete("/folder/:id/delete", ensureAuthenticated, async (req, res) => {
   const { id } = req.params;
-  try{
+  console.log(`Delete request received for folder ID:, ${id}`);
+
+  try {
+    const folder = await prisma.folder.findUnique({
+      where: { id: parseInt(id) },
+      include: { files: true },
+    });
+    console.log("Folder details:", folder);
+
+    if (!folder) {
+      console.log("Folder not found");
+      return res.status(404).json({ error: "folder not found" });
+    }
+
+    console.log("Deleting files associated with folder");
+    await prisma.file.deleteMany({
+      where: { folderId: parseInt(id) },
+    }); 
+    console.log("Files deleted");
+
     await prisma.folder.delete({
       where: { id: parseInt(id) },
     });
-    res.status(204).send();
-  } catch(error){
-    res.status(500).json({ error: "Error deleting folder"});
+    console.log("Redirecting to /dashboard");
+    res.redirect("/dashboard");
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ error: "Error deleting folder" });
   }
-})
+});
 
 module.exports = router;
